+++ src/aks-preview/azext_aks_preview/custom.py
@@ -722,8 +722,6 @@ def aks_create(
     enable_windows_recording_rules=False,
     # azure monitor profile - app monitoring
     enable_azure_monitor_app_monitoring=False,
-    # azure monitor profile - logs
-    enable_azure_monitor_logs=False,
     # opentelemetry parameters
     enable_opentelemetry_metrics=False,
     opentelemetry_metrics_port=None,
@@ -927,9 +925,6 @@ def aks_update(
     # azure monitor profile - app monitoring
     enable_azure_monitor_app_monitoring=False,
     disable_azure_monitor_app_monitoring=False,
-    # azure monitor profile - logs
-    enable_azure_monitor_logs=False,
-    disable_azure_monitor_logs=False,
     # opentelemetry parameters
     enable_opentelemetry_metrics=False,
     opentelemetry_metrics_port=None,